fix some compile issues encountered by interoping with golang

+++ interface/include/texturemanager_bind.h
@@ -4,4 +4,4 @@
 
 DLL TextureManagerHandle texturemanager_singleton();
 
-DLL void texturemanager_set_default_num_mipmaps(TextureManagerHandle tm_hande, size_t number);
+DLL void texturemanager_set_default_num_mipmaps(TextureManagerHandle tm_hande, int number);
