+++ interface/source/texturemanager_bind.cpp
@@ -7,8 +7,8 @@ DLL TextureManagerHandle texturemanager_singleton()
   return reinterpret_cast<TextureManagerHandle>(Ogre::TextureManager::getSingletonPtr());
 }
 
-DLL void texturemanager_set_default_num_mipmaps(TextureManagerHandle tm_hande, size_t number)
+DLL void texturemanager_set_default_num_mipmaps(TextureManagerHandle tm_hande, int number)
 {
   Ogre::TextureManager* tm = reinterpret_cast<Ogre::TextureManager*>(tm_hande);
-  tm->setDefaultNumMipmaps(number);
+  tm->setDefaultNumMipmaps(static_cast<size_t>(number));
 }
